+++ src/main/java/fpt/swp391/GlucoTrackAlert/dto/AssignmentRequest.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 @Setter
 public class AssignmentRequest {
 
-    private Integer doctorId;
+    private Long doctorId;
     private Long patientId;
     private String note;
     private String status;
