+++ src/main/java/fpt/swp391/GlucoTrackAlert/dto/login/LoginResponse.java
@@ -13,5 +13,5 @@ public class LoginResponse {
     private String email;
     private String role;
     private String message;
-    private Integer doctorId;
+    private Long doctorId;
 }
