+++ src/main/java/fpt/swp391/GlucoTrackAlert/service/impl/DoctorPatientAssignmentService.java
@@ -137,7 +137,7 @@ public class DoctorPatientAssignmentService {
         return assignmentRepository.findAll();
     }
 
-    public DoctorPatientAssignment updateAssignment(Integer id, AssignmentRequest request) {
+    public DoctorPatientAssignment updateAssignment(Long id, AssignmentRequest request) {
         DoctorPatientAssignment updatedAssignment = new DoctorPatientAssignment();
         if (request.getDoctorId() != null) {
             Doctor doctor = doctorRepository.findById(request.getDoctorId())
@@ -157,7 +157,7 @@ public class DoctorPatientAssignmentService {
         return updateAssignment(id, updatedAssignment);
     }
 
-    public DoctorPatientAssignment updateAssignment(Integer id, DoctorPatientAssignment updatedAssignment) {
+    public DoctorPatientAssignment updateAssignment(Long id, DoctorPatientAssignment updatedAssignment) {
         DoctorPatientAssignment assignment
                 = assignmentRepository.findById(id)
                         .orElseThrow(() -> new RuntimeException("Assignment not found"));
@@ -196,7 +196,7 @@ public class DoctorPatientAssignmentService {
         return assignmentRepository.save(assignment);
     }
 
-    public void deleteAssignment(Integer id) {
+    public void deleteAssignment(Long id) {
         DoctorPatientAssignment assignment
                 = assignmentRepository.findById(id)
                         .orElseThrow(() -> new RuntimeException("Assignment not found"));
@@ -204,7 +204,7 @@ public class DoctorPatientAssignmentService {
         assignmentRepository.save(assignment);
     }
 
-    public List<Map<String, Object>> getPatientsByDoctor(Integer doctorId) {
+    public List<Map<String, Object>> getPatientsByDoctor(Long doctorId) {
         return assignmentRepository
                 .findByDoctorIdAndStatus(doctorId, "active")
                 .stream()
@@ -224,7 +224,7 @@ public class DoctorPatientAssignmentService {
                 .toList();
     }
 
-    public void hardDeleteAssignment(Integer id) {
+    public void hardDeleteAssignment(Long id) {
         DoctorPatientAssignment assignment
                 = assignmentRepository.findById(id)
                         .orElseThrow(() -> new RuntimeException("Assignment not found"));
